api and handle exception of comment function

+++ src/main/java/com/huism/community/controller/PublishController.java
@@ -31,7 +31,7 @@ public class PublishController {
     public String doPublish(@RequestParam(value = "title",required = false) String title,
                             @RequestParam(value = "description",required = false) String description,
                             @RequestParam(value = "tag",required = false) String tag,
-                            @RequestParam(value = "id",required = false) Integer id,
+                            @RequestParam(value = "id",required = false) Long id,
                             HttpServletRequest request,
                             Model model){
         // 会写表单数据
@@ -81,7 +81,7 @@ public class PublishController {
 
     // 编辑问题请求
     @GetMapping("/publish/{id}")
-    public String edit(@PathVariable(name = "id")Integer id, Model model){
+    public String edit(@PathVariable(name = "id") Long id, Model model){
         QuestionDTO question = questionService.getById(id);
 
         model.addAttribute("title",question.getTitle());
